+++ app/_components/GraphicalSection/Icons/BurningIcon.tsx
