+++ app/_components/GraphicalSection/Icons/CertificateIcon.tsx
