+++ app/_components/GraphicalSection/Icons/MintingIcon.tsx
