+++ app/_components/GraphicalSection/Icons/WithdrawalIcon.tsx
