+++ app/_components/GraphicalSection/Icons/index.ts
@@ -1,0 +1,4 @@
+export { BurningIcon } from "./BurningIcon";
+export { CertificateIcon } from "./CertificateIcon";
+export { MintingIcon } from "./MintingIcon";
+export { WithdrawalIcon } from "./WithdrawalIcon";
