+++ app/_components/GraphicalSection/InfoPanel/AssetCard.tsx
