+++ app/_components/GraphicalSection/InfoPanel/InfoPanel.tsx
