+++ app/_components/GraphicalSection/InfoPanel/JSONModal.tsx
