+++ app/_components/GraphicalSection/InfoPanel/TxInfo.tsx
