+++ app/_components/GraphicalSection/InfoPanel/index.ts
