+++ app/_components/GraphicalSection/index.ts
@@ -1,3 +1,4 @@
 export { UtxoLine as Line } from "./Line";
+export { Playground } from "./Playground";
 export { Transaction } from "./Transaction";
 export { Utxo } from "./Utxo";
