+++ app/_components/index.ts
@@ -6,10 +6,9 @@ export { DissectSection } from "./DissectSection/DissectSection";
 export { DslSection } from "./DslSection/DslSection";
 export { Error } from "./Error";
 export { Examples, ExamplesAddress } from "./Examples";
+export { Line, Playground, Transaction, Utxo } from "./GraphicalSection/";
+export { InfoPanel, TxInfo, UtxoInfo } from "./GraphicalSection/InfoPanel";
 export { Header } from "./Header/Header";
-export { InfoPanel, TxInfo, UtxoInfo } from "./InfoPanel";
 export { AddressInput } from "./Input/AddressInput";
 export { CustomInput as Input } from "./Input/Input";
 export { TxInput } from "./Input/TxInput";
-export { Playground } from "./Playground";
-export { Line, Transaction, Utxo } from "./Transaction";
